changes in database types

+++ SmartFridgeApp.Domain/Models/FoodProducts/Category.cs
@@ -2,7 +2,7 @@
 {
     public class Category
     {
-        public byte CategoryId { get; set; }
+        public short CategoryId { get; set; }
         public string Name { get; set; }
 
         public Category(string name)
